SafeRm

Description

Provides functionality for moving a file to a trash location.

Since: 0.1

Delete

delete :: forall env m. (MonadFsReader m, MonadFsWriter m, HasCallStack, HasTrashHome env, MonadLoggerContext m, MonadCallStack m, MonadReader env m, MonadUnliftIO m, MonadSystemTime m) => UniqueSeq (PathI OriginalPath) -> m ()

delete trash p moves path p to the given trash location trash and writes an entry in the trash index. If the trash location is not given, defaults to ~/.trash.

Since: 0.1

deletePermanently :: forall env m. (MonadFsReader m, MonadFsWriter m, HasCallStack, HasTrashHome env, MonadLoggerContext m, MonadCallStack m, MonadReader env m, MonadUnliftIO m, MonadTerminal m) => Bool -> UniqueSeq (PathI TrashName) -> m ()

Permanently deletes the paths from the trash.

Since: 0.1

emptyTrash :: forall env m. (MonadFsReader m, MonadFsWriter m, HasCallStack, HasTrashHome env, MonadLoggerContext m, MonadIO m, MonadReader env m, MonadTerminal m) => Bool -> m ()

Empties the trash. Deletes the index file.

Since: 0.1

Restore

restore :: forall env m. (MonadFsReader m, MonadFsWriter m, HasCallStack, HasTrashHome env, MonadLoggerContext m, MonadCallStack m, MonadReader env m, MonadUnliftIO m) => UniqueSeq (PathI TrashName) -> m ()

restore trash p restores the trashed path <trash>/p to its original location. If trash is not given then we look in the default location e.g. ~/.trash.

Since: 0.1

Information

getIndex :: forall env m. (MonadFsReader m, HasCallStack, HasTrashHome env, MonadLoggerContext m, MonadCallStack m, MonadIO m, MonadReader env m) => m Index

Reads the index at either the specified or default location. If the file does not exist, returns empty.

Since: 0.1

getMetadata :: forall env m. (MonadFsReader m, HasCallStack, HasTrashHome env, MonadLoggerContext m, MonadCallStack m, MonadIO m, MonadReader env m) => m Metadata

Retrieves metadata for the trash directory.

Since: 0.1