SafeRm.Data.Index

Description

Provides types.

Since: 0.1

Documentation

newtype Index

Index that stores the trash data.

Since: 0.1

Constructors

MkIndex

Fields

unIndex :: HashMap (PathI TrashName) PathData

Reading

readIndex :: (MonadFsReader m, MonadLoggerContext m, MonadCallStack m, MonadIO m) => PathI TrashIndex -> m Index

Attempts to read the trash index file. If successful, guarantees:

All trash paths are unique.

Every index entry corresponds to a path in the trash directory.

Since: 0.1

searchIndex

Arguments

:: UniqueSeq (PathI TrashName)
The top-level trash keys to find e.g. foo for ~/.trash/foo.

-> Index
The trash index.

-> ([PathI TrashName], UniqueSeq PathData)
The trash data matching the input keys.

Searches the trash index for keys.

Since: 0.1

Writing

appendIndex :: MonadFsWriter m => PathI TrashIndex -> Index -> m ()

Appends the path data to the trash index. The header is not included.

Since: 0.1

writeIndex :: MonadFsWriter m => PathI TrashIndex -> Index -> m ()

Writes the path data to the trash index, overwriting the index if it exists. The header is included.

Since: 0.1