SafeRm.Data.Metadata

Description

Provides metadata functionality.

Since: 0.1

Documentation

data Metadata

Holds trash metadata.

Since: 0.1

Constructors

MkMetadata

Fields

numEntries :: !Natural

Number of top level entries in the trash index. This should be the same as the index length.

Since: 0.1

numFiles :: !Natural

Number of total files in the trash.

Since: 0.1

logSize :: !(SomeSize Double)

Size of the log file.

Since: 0.1

size :: !(SomeSize Double)

Total size of the trash directory.

Since: 0.1

toMetadata :: (MonadFsReader m, HasCallStack, MonadLoggerContext m, MonadCallStack m, MonadIO m) => (PathI TrashHome, PathI TrashIndex, PathI TrashLog) -> m Metadata

Returns metadata for the trash directory.

Since: 0.1