SafeRm.Data.PathData

Description

Provides types.

Since: 0.1

Documentation

data PathData

Data for a path.

Since: 0.1

Constructors

MkPathData

Fields

pathType :: !PathType

The type of the path.

Since: 0.1

fileName :: !(PathI TrashName)

The path to be used in the trash directory.

Since: 0.1

originalPath :: !(PathI OriginalPath)

The original path on the file system.

Since: 0.1

created :: !Timestamp

Time this entry was created.

Since: 0.1

Creation

toPathData :: (MonadFsReader m, HasCallStack, MonadCallStack m, MonadIO m) => Timestamp -> PathI TrashHome -> PathI OriginalPath -> m PathData

For a given filepath, attempts to capture the following data:

Canonical path.

Unique name to be used in the trash directory.

File/directory type.

Since: 0.1

mvOriginalToTrash :: (MonadFsWriter m, HasCallStack) => PathI TrashHome -> PathData -> m ()

Moves the PathData's originalPath to the trash.

Since: 0.1

Existence

trashPathExists :: (MonadFsReader m, HasCallStack) => PathI TrashHome -> PathData -> m Bool

Returns True if the PathData's fileName corresponds to a real path that exists in TrashHome.

Since: 0.1

originalPathExists :: (MonadFsReader m, HasCallStack) => PathData -> m Bool

Returns True if the PathData's originalPath corresponds to a real path that exists.

Since: 0.1

Deletion

mvTrashToOriginal :: (MonadFsReader m, MonadFsWriter m, HasCallStack, MonadCallStack m, MonadIO m) => PathI TrashHome -> PathData -> m ()

Moves the PathData's fileName back to its originalPath.

Since: 0.1

deletePathData :: MonadFsWriter m => PathI TrashHome -> PathData -> m ()

Permanently deletes the PathData.

Since: 0.1

Sorting

sortDefault :: PathData -> PathData -> Ordering

Sorts by the created date then the name.

Since: 0.1

sortCreated :: PathData -> PathData -> Ordering

Sorts by the created date.

Since: 0.1

sortName :: PathData -> PathData -> Ordering

Sorts by the name.

Since: 0.1

Miscellaneous

headerNames :: (IsList a, IsString (Item a)) => a

Header names.

Since: 0.1

pathDataToTrashPath :: PathI TrashHome -> PathData -> PathI TrashPath

Gives the PathData's full trash path in the given TrashHome.

Since: 0.1