SafeRm.Data.Paths

Description

Provides functionality for distinguishing path types.

Since: 0.1

Types

newtype PathI (i :: PathIndex)

Indexed FilePath so that we can prevent mixing up different filepaths.

Since: 0.1

Constructors

MkPathI

Fields

unPathI :: FilePath

data PathIndex

Types of filepaths used in SafeRm.

Since: 0.1

Constructors

TrashHome
The trash directory.

Since: 0.1

TrashIndex
The trash index file.

Since: 0.1

TrashName
The name corresponding to some file/directory in the trash directory.

Since: 0.1

OriginalPath
The original path for some file/directory in the trash directory.

Since: 0.1

TrashPath
The full trash path i.e. TrashHome </> TrashName

Since: 0.1

TrashLog
The trash log file.

Since: 0.1

Functions

Specific

indexToHome :: PathI TrashIndex -> PathI TrashHome

Returns the trash index's home directory.

Since: 0.1

General

These functions allows for lifting arbitrary FilePath functions onto our PathI. Note that this can easily invalidate any invariants we would like to hold (e.g. appending a path can turn a directory into a file), so caution must be exercised.

reindex :: PathI i -> PathI j

Modifies the index.

Since: 0.1

(<//>) :: PathI i1 -> PathI i2 -> PathI i3 infixr 5

(</>) lifted to PathI. Notice the index can change, so take care.

Since: 0.1

applyPathI :: HasCallStack => (HasCallStack => FilePath -> a) -> PathI i -> a

Lifts a FilePath function to PathI.

Since: 0.1

liftPathI :: (FilePath -> FilePath) -> PathI i -> PathI j

Lifts a FilePath transformation to PathI, allowing for the index to change.

Since: 0.1

liftPathI' :: (FilePath -> FilePath) -> PathI i -> PathI i

liftPathI specialized to the same index. This should be preferred as the former is easier to use incorrectly.

Since: 0.1

liftPathIF :: (Functor f, HasCallStack) => (HasCallStack => FilePath -> f FilePath) -> PathI i -> f (PathI j)

Lifts an effectful FilePath transformation to PathI.

Since: 0.1

liftPathIF' :: Functor f => (HasCallStack => FilePath -> f FilePath) -> PathI i -> f (PathI i)

liftPathIF specialized to the same index. This should be preferred as the former is easier to use incorrectly.

Since: 0.1