SafeRm.Exception

Description

Provides exceptions used by SafeRm.

Since: 0.1

Types

Specific

data PathNotFoundE

Path is not found.

Since: 0.1

Constructors

MkPathNotFoundE !FilePath !CallStack

data RenameDuplicateE

Could not rename file due to duplicate names.

Since: 0.1

Constructors

MkRenameDuplicateE !(PathI TrashName) !CallStack

data ReadIndexE

Error reading the index.

Since: 0.1

Constructors

MkReadIndexE !(PathI TrashIndex) !String !CallStack

data DuplicateIndexPathE

Duplicate paths found in index file.

Since: 0.1

Constructors

MkDuplicateIndexPathE !(PathI TrashIndex) !(PathI TrashName) !CallStack

data TrashPathNotFoundE

Path not found in trash error.

Since: 0.1

Constructors

MkTrashPathNotFoundE !(PathI TrashHome) !(PathI TrashName) !CallStack

data RestoreCollisionE

Collision with existing file when attempting a restore.

Since: 0.1

Constructors

MkRestoreCollisionE !(PathI TrashName) !(PathI OriginalPath) !CallStack

data IndexSizeMismatchE

Index size did not match the trash directory.

Since: 0.1

Constructors

MkIndexSizeMismatchE !(PathI TrashHome) !Int !Int !CallStack

data TomlDecodeE

Error decoding TOML file.

Since: 0.1

Constructors

MkTomlDecodeE !TOMLError !CallStack

data ArbitraryE

Arbitrary exception. SomeException with CallStack.

Since: 0.1

Constructors

MkArbitraryE !SomeException !CallStack

Aggregating

newtype Exceptions

Aggregates multiple exceptions.

Since: 0.1

Constructors

MkExceptions

Fields

unExceptions :: NonEmpty SomeException

Since: 0.1

Functions

withStackTracing :: (HasCallStack, MonadCallStack m, MonadUnliftIO m) => m a -> m a

Tries the monadic action. If an exception is encountered, it is wrapped in an ArbitraryE and rethrown.

Since: 0.1

displayTraceIf :: Exception e => Bool -> e -> Text

If True, uses displayTrace; otherwise uses displayException.

Since: 0.1

displayTrace :: Exception e => e -> Text

Like displayException, except we attempt to match the exceptions specified in this module, and -- if found -- include the stack trace in the display text. If we do fail to match, simply uses displayException.

Since: 0.1

Optics

_MkPathNotFoundE :: Iso' PathNotFoundE (FilePath, CallStack)

Since: 0.1

_MkRenameDuplicateE :: Iso' RenameDuplicateE (PathI 'TrashName, CallStack)

Since: 0.1

_MkReadIndexE :: Iso' ReadIndexE (PathI 'TrashIndex, String, CallStack)

Since: 0.1

_MkDuplicateIndexPathE :: Iso' DuplicateIndexPathE (PathI 'TrashIndex, PathI 'TrashName, CallStack)

Since: 0.1

_MkTrashPathNotFoundE :: Iso' TrashPathNotFoundE (PathI 'TrashHome, PathI 'TrashName, CallStack)

Since: 0.1

_MkRestoreCollisionE :: Iso' RestoreCollisionE (PathI 'TrashName, PathI 'OriginalPath, CallStack)

Since: 0.1

_MkIndexSizeMismatchE :: Iso' IndexSizeMismatchE (PathI 'TrashHome, Int, Int, CallStack)

Since: 0.1

_MkTomlDecodeE :: Iso' TomlDecodeE (TOMLError, CallStack)

Since: 0.1

_MkArbitraryE :: Iso' ArbitraryE (SomeException, CallStack)

Since: 0.1