SafeRm.Runner

Description

This modules provides an executable for running safe-rm.

Since: 0.1

Main functions

runSafeRm :: (MonadFsReader m, MonadFsWriter m, HasCallStack, MonadCallStack m, MonadUnliftIO m, MonadTerminal m, MonadSystemTime m) => m ()

Entry point for running SafeRm. Does everything: reads CLI args, optional Toml config, and creates the environment before running SafeRm.

Since: 0.1

runCmd :: (MonadFsReader m, MonadFsWriter m, HasCallStack, HasTrashHome env, MonadLoggerContext m, MonadCallStack m, MonadReader env m, MonadUnliftIO m, MonadTerminal m, MonadSystemTime m) => Command -> m ()

Runs SafeRm in the given environment. This is useful in conjunction with getConfiguration as an alternative runSafeRm, when we want to use a custom env.

Helpers

getEnv :: (HasCallStack, MonadFsReader m, MonadFsWriter m, MonadCallStack m, MonadUnliftIO m) => m (Env, Command)

Parses CLI Args and optional TomlConfig to produce the final Env used by SafeRm.

Since: 0.1

getConfiguration :: (HasCallStack, MonadFsReader m, MonadCallStack m, MonadUnliftIO m) => m (TomlConfig, Command)

Parses CLI Args and optional TomlConfig to produce the user configuration. For values shared between the CLI and Toml file, the CLI takes priority.

For example, if both the CLI and Toml file specify the trash home, then the CLI's value will be used.

Since: 0.1