SafeRm.Runner.Args

Description

Provides CLI args functionality.

Since: 0.1

Documentation

getArgs :: IO Args

Retrieves CLI args.

Since: 0.1

data Args

CLI args.

Since: 0.1

Constructors

MkArgs

Fields

tomlConfigPath :: !TomlConfigPath

Path to toml config.

Since: 0.1

trashHome :: !(Maybe (PathI TrashHome))

Path to trash home.

Since: 0.1

logLevel :: !(Maybe (Maybe LogLevel))

The file logging level.

Since: 0.1

showTrace :: !(Maybe Bool)

If enabled, prints the stack trace for any errors encountered.

Since: 0.1

noShowTrace :: !Bool

Disables showTrace. This exists because --showTrace false cannot be specified on the CLI, as it is a simple switch.

Since: 0.1

command :: !Command

Command to run.

Since: 0.1

data TomlConfigPath

Toml path config.

Since: 0.1

Constructors

TomlNone
Do not use any Toml config.

Since: 0.1

TomlDefault
Attempts to read the Toml file at the default path.

Since: 0.1

TomlPath !FilePath
Path to Toml file.

Since: 0.1

_TomlNone :: Prism' TomlConfigPath ()

_TomlDefault :: Prism' TomlConfigPath ()

_TomlPath :: Prism' TomlConfigPath FilePath