SafeRm.Runner.Toml

Description

Provides TOML configuration.

Since: 0.1

Documentation

data TomlConfig

Holds TOML configuration.

Since: 0.1

Constructors

MkTomlConfig

Fields

trashHome :: !(Maybe (PathI TrashHome))

logLevel :: !(Maybe (Maybe LogLevel))

showTrace :: !(Maybe Bool)

mergeConfigs :: Args -> TomlConfig -> TomlConfig

Merges the args and toml config into a single toml config. If some field F is specified by both args and toml config, then args takes precedence.

Since: 0.1